HW-57660 move test to proper folder, add to HyperwalletTestSuite
